Temp-dir repo should periodically purge files.
When using a temporary directory, the content should periodically be cleaned up. As a new random directory is used for each invocation, files should be deleted upon exit.

This commit adds a cleanup task that runs every five minutes. It cleans files from the repository in a FIFO style, until the size of the repository is no longer larger
than the amount of free space available. This intends to safeguard against out-of-discspace situations.